feat: Entity 및 Repository 추가 (#7)
