+++ breaditnow-domain/src/main/java/com/breaditnow/domain/bakery/enumerate/OperatingStatus.java
@@ -1,0 +1,5 @@
+package com.breaditnow.domain.bakery.enumerate;
+
+public enum OperatingStatus {
+	OPEN, CLOSED, TEMPORARY_CLOSED;
+}
