+++ breaditnow-domain/src/main/java/com/breaditnow/domain/customer/enumerate/Provider.java
@@ -1,0 +1,5 @@
+package com.breaditnow.domain.customer.enumerate;
+
+public enum Provider {
+	EMAIL, KAKAO, NAVER, GOOGLE;
+}
