+++ breaditnow-domain/src/main/java/com/breaditnow/domain/customer/repository/CustomerRepository.java
@@ -1,0 +1,10 @@
+package com.breaditnow.domain.customer.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.breaditnow.domain.customer.entity.Customer;
+
+@Repository
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
+}
