+++ breaditnow-domain/src/main/java/com/breaditnow/domain/owner/repository/OwnerRepository.java
@@ -1,0 +1,10 @@
+package com.breaditnow.domain.owner.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.breaditnow.domain.owner.entity.Owner;
+
+@Repository
+public interface OwnerRepository extends JpaRepository<Owner, Long> {
+}
