+++ breaditnow-domain/src/main/java/com/breaditnow/domain/product/enumerate/ProductType.java
@@ -1,0 +1,5 @@
+package com.breaditnow.domain.product.enumerate;
+
+public enum ProductType {
+	BREAD, OTHER;
+}
