+++ breaditnow-domain/src/main/java/com/breaditnow/domain/product/repository/ProductRepository.java
@@ -1,0 +1,8 @@
+package com.breaditnow.domain.product.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.breaditnow.domain.product.entity.Product;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+}
