--- breaditnow-mysql/src/main/java/com/breaditnow/mysql/domain/bakery/enumerate/OperatingStatus.java
@@ -1,5 +1,0 @@
-package com.breaditnow.mysql.domain.bakery.enumerate;
-
-public enum OperatingStatus {
-	OPEN, CLOSED, TEMPORARY_CLOSED;
-}
