--- breaditnow-mysql/src/main/java/com/breaditnow/mysql/domain/customer/enumerate/Provider.java
@@ -1,5 +1,0 @@
-package com.breaditnow.mysql.domain.customer.enumerate;
-
-public enum Provider {
-	EMAIL, KAKAO, NAVER, GOOGLE;
-}
