--- breaditnow-mysql/src/main/java/com/breaditnow/mysql/domain/customer/repository/CustomerRepository.java
@@ -1,10 +1,0 @@
-package com.breaditnow.mysql.domain.customer.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.breaditnow.mysql.domain.customer.entity.Customer;
-
-@Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
-}
