--- breaditnow-mysql/src/main/java/com/breaditnow/mysql/domain/owner/repository/OwnerRepository.java
@@ -1,10 +1,0 @@
-package com.breaditnow.mysql.domain.owner.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.breaditnow.mysql.domain.owner.entity.Owner;
-
-@Repository
-public interface OwnerRepository extends JpaRepository<Owner, Long> {
-}
