--- breaditnow-mysql/src/main/java/com/breaditnow/mysql/domain/product/enumerate/ProductType.java
@@ -1,5 +1,0 @@
-package com.breaditnow.mysql.domain.product.enumerate;
-
-public enum ProductType {
-	BREAD, OTHER;
-}
