--- breaditnow-mysql/src/test/java/com/breaditnow/mysql/domain/customer/entity/CustomerTest.java
@@ -1,15 +1,0 @@
-package com.breaditnow.mysql.domain.customer.entity;
-
-import org.junit.jupiter.api.Test;
-
-class CustomerTest {
-	@Test
-	void builder를_테스트한다() {
-		Customer customer = Customer.createCustomerBuilder()
-			.email("email@naver.com")
-			.password("password1!")
-			.build();
-
-		System.out.println(customer);
-	}
-}
